+++ src/lib/tables.c
@@ -11,7 +11,7 @@ const int slen1_tab[16] = { 0, 0, 0, 0, 3, 1, 1, 1, 2, 2, 2, 3, 3, 3, 4, 4 };
 const int slen2_tab[16] = { 0, 1, 2, 3, 0, 1, 2, 3, 1, 2, 3, 1, 2, 3, 2, 3 };
 
 /* Valid samplerates and bitrates. */
-const long samplerates[9] = {
+const int samplerates[9] = {
   44100, 48000, 32000, /* MPEG-I */
   22050, 24000, 16000, /* MPEG-II */
   11025, 12000, 8000   /* MPEG-2.5 */
