+++ src/lib/tables.h
@@ -6,7 +6,7 @@
 extern const int slen1_tab[16];
 extern const int slen2_tab[16];
 
-extern const long samplerates[9];
+extern const int samplerates[9];
 extern const int bitrates[16][4];
 
 extern const int  shine_scale_fact_band_index[9][23];
